requirements.txt
# Real-Time Malware Detection System Dependencies
# FastAPI + Pydantic v2 + PyTorch 2.0+

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Machine Learning
torch>=2.0.0
numpy>=1.24.0

# HTTP Requests
httpx>=0.26.0  # Async HTTP client for streaming
requests>=2.31.0

# Configuration
pyyaml>=6.0.1

# System Monitoring
psutil>=5.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0  # For test client

# Utilities
python-multipart>=0.0.6  # For file uploads

# Logging
python-json-logger>=2.0.7